requirements.txt
streamlit
pandas
scikit-learn
plotly
numpy
mysql-connector-python